refactor: compose country module traits
