+++ packages/tokens/src/rwa/compliance/modules/country_restrict/mod.rs
@@ -10,9 +10,9 @@ pub mod storage;
 #[cfg(test)]
 mod test;
 
-use soroban_sdk::{contractevent, contracttrait, Address, Env, String, Vec};
+use soroban_sdk::{contractevent, contracttrait, Address, Env, Vec};
 
-use super::storage::{get_compliance_address, module_name};
+use super::ComplianceModule;
 
 /// Emitted when a country is added to the restriction list.
 #[contractevent]
@@ -61,7 +61,7 @@ pub fn emit_country_unrestricted(e: &Env, token: &Address, country: u32) {
 /// no default implementation because each contract must enforce its own access
 /// control before delegating to storage helpers.
 #[contracttrait]
-pub trait CountryRestrict {
+pub trait CountryRestrict: ComplianceModule {
     /// Configures the Identity Registry Storage contract for `token`.
     ///
     /// # Arguments
@@ -164,91 +164,4 @@ pub trait CountryRestrict {
     fn is_country_restricted(e: &Env, token: Address, country: u32) -> bool {
         storage::is_country_restricted(e, &token, country)
     }
-
-    /// Country restriction does not track transfer state.
-    ///
-    /// # Arguments
-    ///
-    /// * `_e` - Access to the Soroban environment.
-    /// * `_from` - The sender address.
-    /// * `_to` - The recipient address.
-    /// * `_amount` - The transfer amount.
-    /// * `_token` - The token address.
-    fn on_transfer(_e: &Env, _from: Address, _to: Address, _amount: i128, _token: Address) {}
-
-    /// Country restriction does not track mint state.
-    ///
-    /// # Arguments
-    ///
-    /// * `_e` - Access to the Soroban environment.
-    /// * `_to` - The recipient address.
-    /// * `_amount` - The minted amount.
-    /// * `_token` - The token address.
-    fn on_created(_e: &Env, _to: Address, _amount: i128, _token: Address) {}
-
-    /// Country restriction does not track burn state.
-    ///
-    /// # Arguments
-    ///
-    /// * `_e` - Access to the Soroban environment.
-    /// * `_from` - The holder address.
-    /// * `_amount` - The burned amount.
-    /// * `_token` - The token address.
-    fn on_destroyed(_e: &Env, _from: Address, _amount: i128, _token: Address) {}
-
-    /// Returns `true` if the transfer recipient has no restricted country.
-    ///
-    /// # Arguments
-    ///
-    /// * `e` - Access to the Soroban environment.
-    /// * `from` - The sender address.
-    /// * `to` - The recipient address.
-    /// * `amount` - The transfer amount.
-    /// * `token` - The token address.
-    fn can_transfer(e: &Env, from: Address, to: Address, amount: i128, token: Address) -> bool {
-        storage::can_transfer(e, &from, &to, amount, &token)
-    }
-
-    /// Returns `true` if the mint recipient has no restricted country.
-    ///
-    /// # Arguments
-    ///
-    /// * `e` - Access to the Soroban environment.
-    /// * `to` - The recipient address.
-    /// * `amount` - The minted amount.
-    /// * `token` - The token address.
-    fn can_create(e: &Env, to: Address, amount: i128, token: Address) -> bool {
-        storage::can_create(e, &to, amount, &token)
-    }
-
-    /// Returns this module's display name.
-    ///
-    /// # Arguments
-    ///
-    /// * `e` - Access to the Soroban environment.
-    fn name(e: &Env) -> String {
-        module_name(e, "CountryRestrictModule")
-    }
-
-    /// Returns the compliance contract address.
-    ///
-    /// # Arguments
-    ///
-    /// * `e` - Access to the Soroban environment.
-    fn get_compliance_address(e: &Env) -> Address {
-        get_compliance_address(e)
-    }
-
-    /// Sets the compliance contract address.
-    ///
-    /// # Arguments
-    ///
-    /// * `e` - Access to the Soroban environment.
-    /// * `compliance` - The compliance contract address.
-    ///
-    /// # Notes
-    ///
-    /// No default implementation is provided because this is a privileged
-    /// bootstrap operation that requires custom access control.
-    fn set_compliance_address(e: &Env, compliance: Address);
 }
